pyproject.toml
[tool.poetry]
name = "jobboard"
version = "0.1.0"
description = ""
authors = ["Kerelko Barabash <kerelkobarabash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-ninja = "^1.1.0"
psycopg2 = "^2.9.9"
factory-boy = "^3.3.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
punq = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[pytest]
DJANGO_SETTINGS_MODULE = "src.core.settings.prod"
